pyproject.toml
[project]
name = "filchat"
version = "0.1.0"
description = "application FilChat"
authors = [
    {name = "Osynia", email = "osynia.devapps@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.10",
    "django-compressor>=4.6.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-tailwind>=4.2.0",
    "flake8-pyproject>=1.2.4",
    "pillow>=12.1.0",
    "pyinstaller>=6.17.0",
    "pyside6>=6.10.1",
    "tailwind>=3.1.5b0",
    "wagtail>=7.2.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "django-debug-toolbar>=6.2.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.black]
line_length = 88
target_version = ['py38']

[tool.flake8]
max-line-length = 120
extend-exclude = ['migrations/', './filchat-0.0/', './.venv/', 'tests/output/', 'settings/']
ignore = ['E402','E501']

[tool.mypy]
